battery/navigator-iframe.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Battery Test: each Navigator object has a battery promise</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<script src="../support/testharness.js"></script>
<script src="../support/testharnessreport.js"></script>
<div id="log"></div>
<iframe src="about:blank"></iframe>
<script>
promise_test(function(t) {
  var iframe = document.querySelector("iframe");
  var originalPromise = navigator.getBattery();

  return originalPromise.then(function(originalManager) {
    var promise = iframe.contentWindow.navigator.getBattery();

    assert_true(originalManager instanceof BatteryManager);
    assert_not_equals(iframe.contentWindow.navigator,
        navigator,
        'navigator objects shall be different');
    assert_not_equals(promise,
        originalPromise,
        'battery status promises shall be different');

    return promise;
  }).then(function(manager) {
    assert_equals(manager.__proto__,
        iframe.contentWindow.BatteryManager.prototype);
    assert_true(manager instanceof iframe.contentWindow.BatteryManager);
  });

}, document.title + ', iframe');
</script>
